+++ Web_Api/Models/Items/ItemExchangeViewModel.cs
@@ -9,6 +9,6 @@ namespace WebApi.Models.Items
     {
         public int ItemId { get; set; }
         public string NewOwnerAccountUserName { get; set; }
-        public int PointValue { get; set; }
+        public double PointValue { get; set; }
     }
 }
